♻️ extract record/assembly
