+++ packages/rum/src/domain/record/observers/mouseInteractionObserver.ts
@@ -3,9 +3,9 @@ import { assign, addEventListeners, DOM_EVENT } from '@datadog/browser-core'
 import { NodePrivacyLevel } from '../../../constants'
 import type { MouseInteraction, MouseInteractionData, BrowserIncrementalSnapshotRecord } from '../../../types'
 import { IncrementalSource, MouseInteractionType } from '../../../types'
+import { assembleIncrementalSnapshot } from '../assembly'
 import { getNodePrivacyLevel } from '../privacy'
 import { getSerializedNodeId, hasSerializedNode } from '../serialization'
-import { assembleIncrementalSnapshot } from '../utils'
 import { tryToComputeCoordinates } from './moveObserver'
 import type { RecordIds } from './recordIds'
 import { getEventTarget } from './utils'
