+++ packages/rum/src/domain/record/utils.ts
@@ -1,23 +1,3 @@
-import { assign, timeStampNow } from '@datadog/browser-core'
-import type { BrowserIncrementalData, BrowserIncrementalSnapshotRecord } from '../../types'
-import { RecordType } from '../../types'
-
 export function isTouchEvent(event: MouseEvent | TouchEvent): event is TouchEvent {
   return Boolean((event as TouchEvent).changedTouches)
-}
-
-export function assembleIncrementalSnapshot<Data extends BrowserIncrementalData>(
-  source: Data['source'],
-  data: Omit<Data, 'source'>
-): BrowserIncrementalSnapshotRecord {
-  return {
-    data: assign(
-      {
-        source,
-      },
-      data
-    ) as Data,
-    type: RecordType.IncrementalSnapshot,
-    timestamp: timeStampNow(),
-  }
 }
